conf files: add web indexes
build: 
-implement more states: timeout testfailure
-add SIGINT handler to allow clean shutdown (unlocks build env)
-add "terminate if no output in x seconds" option

do-build: execute as the correct user

monobuild:
-web updates
 -skip empty platforms that have no packages
 -switch columns and rows
 -consolidate noarch packages into one row
 -make use of new states

In order to "terminate if no output in x seconds", process handling had to 
be made non-blocking, which also requires some minor logging updates

launch-process.py: will be used to implement output_terminate on client side
-this will help shut things down in the build env better

pyutil files:
-changes to implement above features







svn path=/trunk/release/; revision=59241

+++ pyutils/packaging.py
@@ -133,7 +133,10 @@ class buildenv:
 		fd.close()
 
 	def unlock_env(self):
-		os.unlink(self.lock_filename)
+		if os.path.exists(self.lock_filename):
+			os.unlink(self.lock_filename)
+		else:
+			print "Build environment already unlocked"
 
 	def is_locked(self):
 		return os.path.exists(self.lock_filename)
@@ -182,8 +185,9 @@ class package:
 				file_string += line
 		def_file.close()
 
+		# TODO: abstract out this bash parsing stuff so it's uniform between conf and def files
 		# get vars (var="value")
-		for match in re.compile('^(\w*?)=[^\(](.*?)"$', re.S | re.M).finditer(file_string):
+		for match in re.compile('^(\w*?)=[^\(](.*?)["\']?$', re.S | re.M).finditer(file_string):
 			self.info[match.group(1)] = match.group(2)
 
 		# get arrays ( var=(value list) )
